# docs/tenant-openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/support/tickets": {
      "post": {
        "operationId": "SupportController_createTicket",
        "summary": "Create support ticket",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTicketDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Support ticket created"
          },
          "400": {
            "description": "Bad Request – validation failed for payload or query parameters"
          },
          "401": {
            "description": "Unauthorized – missing or invalid JWT"
          },
          "403": {
            "description": "Forbidden – insufficient role or tenant scope"
          }
        },
        "tags": [
          "Support"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "get": {
        "operationId": "SupportController_getAllTickets",
        "summary": "Get all tickets (Admin only)",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-indexed)",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search query string",
            "schema": {
              "example": "revenue",
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Sort field and direction (e.g., createdAt:desc, name:asc)",
            "schema": {
              "example": "createdAt:desc",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "All tenant tickets"
          },
          "400": {
            "description": "Bad Request – validation failed for payload or query parameters"
          },
          "401": {
            "description": "Unauthorized – missing or invalid JWT"
          },
          "403": {
            "description": "Forbidden – insufficient role or tenant scope"
          }
        },
        "tags": [
          "Support"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/support/tickets/my": {
      "get": {
        "operationId": "SupportController_getMyTickets",
        "summary": "Get my tickets",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-indexed)",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search query string",
            "schema": {
              "example": "revenue",
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Sort field and direction (e.g., createdAt:desc, name:asc)",
            "schema": {
              "example": "createdAt:desc",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tickets created by current user"
          },
          "400": {
            "description": "Bad Request – validation failed for payload or query parameters"
          },
          "401": {
            "description": "Unauthorized – missing or invalid JWT"
          },
          "403": {
            "description": "Forbidden – insufficient role or tenant scope"
          }
        },
        "tags": [
          "Support"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/support/tickets/{id}": {
      "patch": {
        "operationId": "SupportController_updateTicketStatus",
        "summary": "Update ticket status (Admin only)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTicketDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated ticket"
          },
          "400": {
            "description": "Bad Request – validation failed for payload or query parameters"
          },
          "401": {
            "description": "Unauthorized – missing or invalid JWT"
          },
          "403": {
            "description": "Forbidden – insufficient role or tenant scope"
          }
        },
        "tags": [
          "Support"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "SupportController_deleteTicket",
        "summary": "Delete ticket (Admin only)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ticket deleted"
          },
          "400": {
            "description": "Bad Request – validation failed for payload or query parameters"
          },
          "401": {
            "description": "Unauthorized – missing or invalid JWT"
          },
          "403": {
            "description": "Forbidden – insufficient role or tenant scope"
          }
        },
        "tags": [
          "Support"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "BG Accountability Tenant API",
    "description": "REST API for the BG Accountability Platform - Tenant Services.\n\n## Overview\nThis API serves both the Tenant Web Dashboard (Next.js) and Mobile App (Flutter).\n\n## Authentication\nAll protected endpoints require a valid JWT Bearer token.\n- Login via `POST /api/v1/auth/login`\n- Include token in header: `Authorization: Bearer <token>`\n- Refresh tokens via `POST /api/v1/auth/refresh`\n\n## Multi-Factor Authentication (MFA)\n- If MFA is enabled, login returns a `tempToken`\n- Complete MFA via `POST /api/v1/auth/mfa/login` with tempToken + TOTP code\n\n## Rate Limiting\n- Default: 100 requests per minute per IP\n- Auth endpoints: 10 requests per minute per IP\n\n## Error Format\nAll errors follow: `{ success: false, message: string, code?: string, details?: object }`\n\n## Pagination\nPaginated endpoints use: `{ items: T[], total: number, page: number, pageSize: number }`\n",
    "version": "1.0.0",
    "contact": {
      "name": "BG Accountability",
      "url": "https://bridgegaps.app",
      "email": "support@bridgegaps.app"
    },
    "license": {
      "name": "Proprietary",
      "url": ""
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3002",
      "description": "Local Development"
    },
    {
      "url": "https://api.bridgegaps.app",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT access token"
      }
    },
    "schemas": {
...
      "CreateTicketDto": {
        "type": "object",
        "properties": {
          "subject": {
            "type": "string",
            "description": "Brief subject line for the ticket",
            "minLength": 5,
            "maxLength": 120,
            "example": "Unable to access dashboard metrics"
          },
          "message": {
            "type": "string",
            "description": "Detailed description of the issue",
            "minLength": 10,
            "maxLength": 2000,
            "example": "When I try to load the metrics page, it shows a blank screen. This started happening after the latest update."
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ],
            "default": "MEDIUM",
            "description": "Ticket priority level"
          }
        },
        "required": [
          "subject",
          "message"
        ]
      },
      "UpdateTicketDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "IN_PROGRESS",
              "RESOLVED",
              "CLOSED"
            ]
          },
          "adminNote": {
            "type": "string",
            "maxLength": 1000
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}